refactor: 초대 관련해서 사용하던 에러 모달을 공용화
- 모달에 들어가는 props이 title, description, buttonText 정도로 공용화 가능하다고 판단하여 전환
- 에러 모달 이름을 변경하면서 해당 모달을 사용하는 파일 수정
