+++ hooks/invite-member/useShareEditLink.ts
@@ -1,7 +1,7 @@
 import { useCallback } from 'react';
 import { useGetOrRefreshEditToken } from './useGetOrRefreshEditToken';
 import { useModalStore } from '@/lib/store/modalStore';
-import { RESOURCE_ROUTE } from '@/lib/constants/inviteResourceType';
+import { RESOURCE_ROUTE } from '@/lib/constants/ResourceType';
 import { ResourceType } from '@/types/invite';
 
 export const useShareEditLink = () => {
@@ -18,7 +18,7 @@ export const useShareEditLink = () => {
             if (!baseUrl) {
               console.error('NEXT_PUBLIC_BASE_URL이 설정되지 않았습니다.');
               open({
-                type: 'inviteError',
+                type: 'error',
                 props: {
                   title: '초대 링크 생성 실패',
                   description: '초대 링크를 생성하는 도중 문제가 발생했습니다. \n 잠시 후 다시 시도해주세요.',
@@ -32,7 +32,7 @@ export const useShareEditLink = () => {
           onError: (error) => {
             console.error(error);
             open({
-              type: 'inviteError',
+              type: 'error',
               props: {
                 title: '초대 링크 생성 실패',
                 description: '초대 링크 생성에 실패했습니다.\n잠시 후 다시 시도해주세요.',
